Move assets to separate files
