+++ lib/assets/db_connection/main.js
@@ -1,0 +1,85 @@
+export function init(ctx, info) {
+  ctx.importCSS("main.css");
+  ctx.importCSS("https://fonts.googleapis.com/css2?family=Inter:wght@400;500&display=swap");
+
+  ctx.root.innerHTML = `
+    <div class="app">
+      <div id="info-box" class="info-box"></div>
+      <div>
+        <div class="row">
+          <input class="input input--xs" name="variable" type="text" />
+          <div>=</div>
+          <select class="input input--sm" name="type">
+            <option value="postgres">PostgreSQL</option>
+            <option value="mysql">MySQL</option>
+          </select>
+        </div>
+        <div class="row">
+          <div class="field grow">
+            <label class="input-label">Hostname</label>
+            <input class="input" name="hostname" type="text" />
+          </div>
+          <div class="field">
+            <label class="input-label">Port</label>
+            <input class="input input--xs" name="port" type="number" />
+          </div>
+        </div>
+        <div class="row">
+          <div class="field grow">
+            <label class="input-label">User</label>
+            <input class="input" name="username" type="text" />
+          </div>
+        </div>
+        <div class="row">
+          <div class="field grow">
+            <label class="input-label">Password</label>
+            <input class="input" name="password" type="password" />
+          </div>
+        </div>
+        <div class="row">
+          <div class="field grow">
+            <label class="input-label">Database</label>
+            <input class="input" name="database" type="text" />
+          </div>
+        </div>
+      </div>
+    </div>
+  `;
+
+  setValues(info.fields);
+
+  updateInfoBox(info.missing_dep);
+
+  ctx.root.addEventListener("blur", handleChange, true);
+  ctx.root.addEventListener("change", handleChange);
+
+  function handleChange(event) {
+    const { value, name } = event.target;
+    ctx.pushEvent("update", { field: name, value });
+  }
+
+  ctx.handleEvent("update", ({ fields }) => {
+    setValues(fields);
+  });
+
+  ctx.handleEvent("missing_dep", ({ dep }) => {
+    updateInfoBox(dep);
+  });
+
+  function setValues(fields) {
+    for (const field in fields) {
+      ctx.root.querySelector(`[name="${field}"]`).value = fields[field];
+    }
+  }
+
+  function updateInfoBox(dep) {
+    const infoBox = ctx.root.querySelector("#info-box");
+
+    if (dep) {
+      infoBox.classList.remove("hidden");
+      infoBox.textContent = `To successfully connect, you need to add the following dependency:\n\n    ${dep}`;
+    } else {
+      infoBox.classList.add("hidden");
+    }
+  }
+}
